Fix TfPathUtils_Python on Windows
On Windows, TfRealPath explicitly lower-cases drive letters for
paths with symlinks(*) while os.path.abspath seems to always
return upper-cased drive letters. This caused the test to
fail.

This has likely been a long-standing failure but had not
been noticed because creating symlinks on Windows with
Python 3.7 or lower required running in a command prompt
as administrator and our test builds just run in a normal
command prompt, so the symlink tests would be skipped entirely.
However, Python 3.8 added support for "unelevated symlinks
on Windows with Developer Mode", which is exactly where we
first started noticing these test failures.

(*) We may want to enforce a consistent policy for all
paths in TfRealPath, but that's a bigger and riskier
change so we just update the test to exercise the current
behavior for now.

(Internal change: 2175812)