I am modifying functions in /arch so they can substitute in a reason or message string.
I am also renaming and adjusting the functionality of ArchLogPostMortem() and
related functions for clarity.  ArchSetPostMortem() is now
ArchSetProcessStateLogCommand(), and takes in two arrays of arguments for fatal
and nonfatal cases, similar to ArchSetLogSession(). The nonfatal case can be
invoked through the new ArchLogCurrentProcessState() and is intended to preserve
the old crash handling behavior of taking a stack trace and only outputting
information to the terminal. The fatal case can be invoked through
ArchLogFatalProcessState(), and is intened to be the normal way of handling a
crash. All code with the old ArchLogPostMortem() has been updated accordingly to
use the appropriate new function.

(Internal change: 2233042)