Resync prim path on connection changed
Fixes #2013

A connection changed notification implies a shader rebuild, which should
force a rescan of cached data for the affected material/light/camera.

In the cas of a material, this will also force a rescan of the dependent
geometries, which is exactly what is needed to fix the refresh issue.