tf: avoid wrapping TfFunctionRef in TfFunctionRef
Because TfFunctionRef has a templated constructor that accepts a
non-const reference, that constructor is preferred over the copy
constructor that takes a const reference when constructing a
TfFunctionRef from a non-const TfFunctionRef value.  The result is a
double indirection when invoking the "copy".  Assignment behaves in
the same way.

Provide an additional copy constructor and assignment operator with
non-const TfFunctionRef arguments that will be selected over the
generic constructor.

use template metaprogramming to ensure TfFunctionRef copies

MSVC does not accept multiple defaulted special member functions so
TfFunctionRef now uses enable_if to disable the generic constructor
and assignment operator when passed a reference to a TfFunctionRef.

(Internal change: 2248831)
(Internal change: 2248954)