tf: Rework mallocTags implementation to be much more concurrency-scalable.  The previous implementation had a global lock around all malloc/free operations which would dramatically slow down parallel applications when tagging was enabled.  This new implementation uses concurrent containers and a more scalable readers-writer lock to speed up tagging concurrent programs.  For example, with malloc tags disabled, it takes about 1 second to open the City_set on a UsdStage. Prior to this change, with malloc tags enabled, it took ~40 seconds. With this change it now takes ~5 seconds.
Enhance the TfAutoMallocTag class to accept any number of tag names
to push and pop.  This obsoletes TfAutoMallocTag2, which remains as a
compatibility alias for TfAutoMallocTag.

Remove the ptmalloc-specific tagging implementation, since we have
not used ptmalloc for years, and with newer glibc's removing the malloc
hooks, we will require our custom allocator to do tagging at all.

Enable the TfMallocTag unit test, which was (unintentionally?) disabled
when not running on ptmalloc3, which has not been the case for many
years.  This revealed a few test cases that were invalid, since they
assume that a newly created thread would inherit the tag stack from the
creating thread.  That's not the case, so I've commented those out for
now.

(Internal change: 2251856)