tf: remove Tf_InitEnvSettings
In change 1937207, Tf_EnvSettingRegistry::Define was made to
accomodate concurrent initialization of TfEnvSetting values and, as a
consequence, Tf_InitializeEnvSetting now also gracefully handles
re-entrance via registry function subscription.  We no longer need to
ensure that the registry instance is constructed before checking that
_value has been set.

The call to Tf_InitEnvSettings on every access was both opaque to the
compiler and resulted in redundant Tf_EnvSettingRegistry
is-constructed checks.

(Internal change: 2259786)