tf: static TLS for malloc tag data
Starting with glibc 2.25, dynamic TLS is allocated using malloc rather
than __libc_memalign.  Because TfMallocTag itself uses TLS in its
malloc implementation, this can result in unbounded recursion.  To
avoid this, the per-thread data pointer should be put into static TLS.
The _ThreadData object itself is not placed in TLS because libtf is
loaded with dlopen in some workflows and the amount of storage
available is limited.

(Internal change: 2333481)