tf: Use capsules to hold TfRefPtr in Python
Tf would previously generate Python bindings for
Tf_PyOwnershipHelper<Ptr, ...>::_RefPtrHolder on demand so it
could pass instances of that object -- and ownership of the
TfRefPtr it holds -- to Python.

Instead of wrapping a C++ class for every RefPtrHolder, we now use
Python capsules to hold heap-allocated RefPtr instances instead.
Capsules can hold on to arbitrary values that are opaque to
Python but can be retrieved by extension modules. This is
fine for our purposes because Python clients don't interact
with these RefPtr objects directly.

(Internal change: 2334588)