OpenUSD: Replace boost::python with pxr/external/boost/python
This change updates all code in the OpenUSD repository to use the
internal pxr/external/boost/python library. The changes here were
largely automated:

- Includes of <boost/python...> headers were changed to
  "pxr/external/boost/python..."

- References to the boost::python namespace were changed to
  pxr_boost:python. For example, "using namespace boost::python"
  statements were updated to "using namespace pxr_boost::python".

  This was mostly a simple search and replace, but some code needed
  to be reordered so that the updated references to pxr_boost::python
  occurred within the context of the open source "pxr" namespace. For
  example, the "using namespace pxr_boost::python" statements
  had to be reordered to come after PXR_NAMESPACE_USING_DIRECTIVE
  in most of the wrap*.cpp files.

- Uses of BOOST_PYTHON_... macros were changed to PXR_BOOST_PYTHON_...

This library is currently a pass-through for boost::python, so there
should be no changes in behavior. Users must still supply boost::python
as a dependency to the build, and any third-party Python bindings that
use boost::python will continue to interoperate with Python-wrapped
types from OpenUSD.

For development purposes, setting PXR_USE_BOOST_PYTHON=OFF when running
cmake will build pxr/external/boost/python in full and remove the
boost::python dependency entirely. This is only a staging measure to
help with development and testing. There are no plans to support this
long-term, and it will be removed when work is complete.

(Internal change: 2339976)