Include export/import macro in `staticTokens.h` documentation
Most usages of the `TF_DECLARE_PUBLIC_TOKENS` macro require symbols to be exported, but the example in the doxygen docs uses the two argument version that doesn't explicitly export symbols.  The example has been updated to use the three argument version that exports symbols with a comment noting that if you don't need to export symbols, how to use the two argument version.

Fixes #3779

Closes #3791

(Internal change: 2378356)